requirements.txt
vanna>=2.0.0
fastapi
uvicorn[standard]
plotly
pandas
python-dotenv
groq
openai
